Show generic wireframe over camera
